+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Requake - Repeating earthquakes search and analysis
 (c) 2021 Claudio Satriano <satriano@ipgp.fr>
 
+## unreleased
+- `plot_families`: print number of events and station-event distance
+
 ## v0.1 - 2021-09-16
 - Initial release, not yet feature complete